+++ events.js
@@ -1,0 +1,4 @@
+export let Open = 'open'
+export let Close = 'close'
+export let Message = 'message'
+export let Error = 'error'