+++ index.js
@@ -1,105 +1,128 @@
 import {createNanoEvents} from 'nanoevents'
+import * as events from './events.js'
 
 
-export const WSEvents = {
-  Open: 'open',
-  Close: 'close',
-  Message: 'message',
-  Error: 'error',
-}
+let CloseCodeNormal = 1000
+let DefaultRetryBackoffInterval = 10
+let DefaultRetryBackoffRate = 2
+let DefaultRetryWaitMax = 10240 // 10 retries
+let DefaultRetryMax = Number.MAX_SAFE_INTEGER
 
-const CloseCodeNormal = 1000
-const DefaultReconnectWait = {
-  Min: 125,
-  Max: 8000,
-}
+
+export {events}
 
 export class WSClient {
-  connected = false
-
-  _emitter = createNanoEvents()
-  _cfg = {
-    url: undefined,
-    protocols: undefined,
+  #connected = false
+  #emitter = createNanoEvents()
+  #cfg = {
     reconnect: true,
-    reconnectWaitMin: DefaultReconnectWait.Min,
-    reconnectWaitMax: DefaultReconnectWait.Max,
-    WebSocket: globalThis?.WebSocket,
+    retryBackoffInterval: DefaultRetryBackoffInterval,
+    retryBackoffRate: DefaultRetryBackoffRate,
+    retryWaitMax: DefaultRetryWaitMax,
+    retryMax: DefaultRetryMax,
+    Websocket: globalThis?.WebSocket,
+  }
+  #ws
+  #retryNo = 0
+  #ready
+  #readyResolve
+
+  get connected() {
+    return this.#connected
+  }
+
+  get ready() {
+    return this.#ready
   }
 
   constructor(cfg) {
-    Object.assign(this._cfg, cfg)
+    Object.assign(this.#cfg, cfg)
 
-    if (!this._cfg.url) {
-      throw new Error("url required")
-    }
+    this.#cfg.url || "url required"()
+    this.#cfg.WebSocket || "WebSocket API not found"()
 
-    this._readyInit()
-    this._reconnectWait = this._cfg.reconnectWaitMin
+    this.#readyInit()
   }
 
-  connect = () => {
-    if (this.connected) return
+  connect = async () => {
+    if (this.connected) {
+      return this
+    }
 
-    this._ws = new this._cfg.WebSocket(this._cfg.url, this._cfg.protocols)
-    this._ws.addEventListener(WSEvents.Open, this._open)
-    this._ws.addEventListener(WSEvents.Close, this._reconnect)
-    this._ws.addEventListener(WSEvents.Message, this._msg)
-    this._ws.addEventListener(WSEvents.Error, this._error)
+    this.#ws = new this.#cfg.WebSocket(this.#cfg.url, this.#cfg.protocols)
+    this.#ws.addEventListener(events.Open, this.#open)
+    this.#ws.addEventListener(events.Close, this.#reconnect)
+    this.#ws.addEventListener(events.Message, this.#msg)
+    this.#ws.addEventListener(events.Error, this.#error)
 
-    return this.ready
+    await this.#ready
+    return this
   }
 
   close(reason) {
-    if (!this.connected) return
-    this._ws.close(CloseCodeNormal, reason)
+    if (!this.#connected) {
+      return
+    }
+
+    this.#ws.close(CloseCodeNormal, reason)
   }
 
-  send(msg) {
-    this.ready.then(() => {
-      this._ws.send(msg)
-    })
+  async send(msg) {
+    await this.#ready
+    this.#ws.send(msg)
   }
 
   on(event, cb) {
-    this._emitter.on(event, cb)
+    return this.#emitter.on(event, cb)
   }
 
 
-  _readyInit = () => {
-    this.ready = new Promise(resolve => this._readyResolve = resolve)
+  #readyInit() {
+    this.#ready = new Promise(resolve => this.#readyResolve = resolve)
   }
 
-  _msg = event => {
-    this._emitter.emit(WSEvents.Message, event.data)
+  #open = event => {
+    this.#connected = true
+    this.#retryNo = 0
+    this.#emitter.emit(events.Open, event)
+    this.#readyResolve()
   }
 
-  _open = event => {
-    this.connected = true
-    this._reconnectWait = this._cfg.reconnectWaitMin
-    this._emitter.emit(WSEvents.Open, event)
-    this._readyResolve()
-  }
-
-  _reconnect = event => {
+  #reconnect = event => {
     if (this.connected) {
       this.connected = false
-      this._readyInit()
-      this._emitter.emit(WSEvents.Close, event)
+      this.#readyInit()
+      this.#emitter.emit(events.Close, event)
     }
 
-    if (event.code === CloseCodeNormal) return
-    if (!this._cfg.reconnect) return
-
-    setTimeout(this.connect, this._reconnectWait)
-    if ((this._reconnectWait *= 2) > this._cfg.reconnectWaitMax) {
-      this._reconnectWait = this._cfg.reconnectWaitMax
+    if (event.code === CloseCodeNormal) {
+      return
     }
+
+    if (!this.#cfg.reconnect) {
+      return
+    }
+
+    if (this.#retryNo >= this.#cfg.retryMax) {
+      return
+    }
+
+    let wait = this.#retryWait()
+    setTimeout(this.connect, wait)
   }
 
-  _error = event => {
-    this._emitter.emit(WSEvents.Error, event)
+  #retryWait() {
+    let {retryBackoffInterval, retryBackoffRate, retryWaitMax} = this.#cfg
+
+    let wait = retryBackoffInterval * retryBackoffRate ** this.#retryNo++
+    return Math.min(wait, retryWaitMax)
+  }
+
+  #msg = event => {
+    this.#emitter.emit(events.Message, event.data)
+  }
+
+  #error = event => {
+    this.#emitter.emit(events.Error, event)
   }
 }
-
-export default WSClient